cleanup of code and some semplifications

+++ src/rendering/solid.vert
@@ -7,11 +7,11 @@ layout(location=2) in vec4 uv_coords;
 layout(location=0) out vec2 v_uv_coords;
 layout(location=1) out vec3 v_n_vector;
 layout(location=2) out int object_id;
-layout(location=3) out vec2 mouse_pos;
+layout(location=3) out ivec2 mouse_pos;
 
 layout(set = 0, binding = 0) uniform Uniforms {
     mat4 u_view_proj;
-    vec2 u_mouse_pos;
+    ivec2 u_mouse_pos;
     vec2 _padding;
 };
